requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
pypdf
tiktoken